fix(nvim): filter out eslint notifications from ALE in noice
ALE is used to run eslint in parallel with the LSP for performance.
However, this was causing annoying notifications to appear in noice.

This commit adds a new filter to noice to specifically ignore
notifications from ALE that contain the word "eslint".

This allows ALE to continue running eslint for linting and fixing,
while preventing the popups from appearing.